feat: add validarUsuarioFlitro.

+++ resolvers/Mutation/usuario.js
@@ -71,12 +71,12 @@ const mutations = {
     },
     async alterarUsuario(_, { filtro, dados }, ctx) {
         ctx && ctx.validaUsuarioFiltro(filtro)
-        
+
         try {
             const usuario = await obterUsuario(_, { filtro })
             if(usuario) {
                 const { id } = usuario
-                if(dados.perfis) {
+                if(ctx.admin && dados.perfis) {
                     await db('usuarios_perfis')
                         .where({ usuario_id: id }).delete()
